Fix setting TRILINOS_BUILD_SHARED_LIBS (CMake 3.1 policy)
It is not clear why this varaible was defined in the first place but with
CMake 3.1.1, this was generating the warning shown below.

I looked where this variable was used in Trilinos and it seems it was only
used in the Mesquite package.  This was getting set with quoted
"@BUILD_SHARED_LIBS@".  I would not have thought that would have dereferenced
the variable BUILD_SHARED_LIBS but apparently it did!.  I have fixed this to
evaluate the varaible properly.

CMake Warning (dev) at Trilinos/cmake/CallbackSetupExtraOptions.cmake:44 (SET):
  Policy CMP0053 is not set: Simplify variable reference and escape sequence
  evaluation.  Run "cmake --help-policy CMP0053" for policy details.  Use the
  cmake_policy command to set the policy and suppress this warning.

  For input:

    '@BUILD_SHARED_LIBS@'

  the old evaluation rules produce:

    'ON'

  but the new evaluation rules produce:

    '@BUILD_SHARED_LIBS@'

  Using the old result for compatibility since the policy is not set.
Call Stack (most recent call first):
  cmake/tribits/core/package_arch/TribitsGlobalMacros.cmake:757 (TRIBITS_REPOSITORY_SETUP_EXTRA_OPTIONS)
  cmake/tribits/core/package_arch/TribitsGlobalMacros.cmake:888 (TRIBITS_REPOSITORY_SETUP_EXTRA_OPTIONS_RUNNER)
  cmake/tribits/core/package_arch/TribitsProjectImpl.cmake:135 (TRIBITS_PROCESS_EXTRA_REPOS_OPTIONS_FILES)
  cmake/tribits/core/package_arch/TribitsProject.cmake:93 (TRIBITS_PROJECT_IMPL)
  CMakeLists.txt:57 (TRIBITS_PROJECT)
This warning is for project developers.  Use -Wno-dev to suppress it.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=712,notpassed=0 (66.50 min)
1) SERIAL_RELEASE => passed: passed=629,notpassed=0 (23.67 min)