Panzer: Changes to DOF support distributed parameter sensitivities
Adding more error reporting to Gather.

Changes to DOF is to allow it to synchronize with the Gather for
distributed parameters. That is a "Fast" DOF computation should not be used
if sensitivities are required only for a distrbuted parameter (not
for the state field). Otherwise there is a seg fault where invalid
memory is accessed.