PyTrilinos: Fix NOX::Petsc compilation errors
References to any classes in the NOX:Petsc namespace within PyTrilinos
are protected with a HAVE_NOX_PETSC macro.  To set this properly for
SWIG, I need to both #include and %include "PyTrilinos_config.h" in
the proper context in order for the generated wrapper codes to
compile.  This commit fixes various SWIG interface files to do this
properly.