Kokkos: Work around tpetra code that tries to attach texture objects to rutime unmanged views.
So I have finally tracked down this error.

It results from using a runtime unmanaged view with a cuda texture
object.  The previous memory management scheme searched the memory_space
for an allocation that bounded the pointer and attached a texture object
to that entity.  This was making the dangerous assumption that the
memory_space initially allocated the unmanaged pointer.  With the new
allocation tracker, it is not possible to bind an unmanaged view (static
or runtime) to a texture object, hence the memory access error.  The
correct way to fix this would be to track down where the runtime
unmanaged views are being constructed, and construct them instead with
an AllocationTracker based on an unmanaged allocator (as seen in the
example code below).

In the short term, I’m checking in a change that will force const,
random read views on cuda to use the __ldg intrinsic, which always work
(given a pointer in the memory space), but is slower (currently) than
tex1Dfetch.

 Example code:

 Instead of creating a runtime unmanaged view

   Kokkos::View<…> view(ptr,…);

Create a managed view base on an unmanaged allocator, i.e., the
allocator’s deallocate method does nothing.

   Impl::AllocationTracker tracker(UnmanagedAllocator,ptr,size,label);
   Kokkos::View<…> view(tracker,…);