Tpetra::MultiVector::elementWiseMultiply: Fixed local kernel.
There was a bug in the local Kokkos::parallel_for functor that
implements Tpetra::MultiVector::elementWiseMultiply.  It didn't
implement the BLAS overwrite rules correctly for scalarThis == 0.
(This corresponds to the beta == 0 case in DGEMV.)  In that case, the
method is supposed to ignore and overwrite the entries of *this.  The
implementation wasn't doing that; it just has a generic kernel that
did the same thing for all values of scalarAB and scalarThis.  As a
result, when the entries of *this were uninitialized, NaNs could end
up in the result.

This issue wasn't manifesting because the new (Kokkos refactor)
version of Tpetra::MultiVector used to ignore the 'zeroOut' argument
to its constructor, that always filled new MultiVector instances with
zeros.  I recently fixed this in the following commits:

a15628a39ce6cdff87e92518bacfd783a5728b95
52cc3fa0849bdcce9d9019096084509cfabbeef0

This fix allowed the elementWiseMultiply bug to manifest.  It appeared
in Ifpack2 and MueLu tests, because Ifpack2's Jacobi (that lives in
Ifpack2::Relaxation) uses elementWiseMultiply to implement the first
Jacobi sweep.

There may be other local kernels that have this issue; I have not
examined them all.  Also, my fix introduces a lot of branches into the
elementWiseMultiply local kernel.  This might reduce performance.  The
way to fix that is to specialize the kernel for the relevant special
cases of scalarAB and scalarThis.  Unless there is pressing need, I'll
wait to do that until we can finally start using C++11 in Tpetra; that
will make the specializations much briefer.

Build/Test Cases Summary
Enabled Packages: Tpetra, Ifpack2, Belos
Disabled Packages: FEI,Moertel,SEACAS,STK,PyTrilinos,CTrilinos,Optika,GlobiPack,OptiPack,RBGen,Rythmos,MOOCHO,NOX,Piro,Sundance,Didasko,FEApp
Enabled all Forward Packages
0) MPI_DEBUG => passed: configure-only passed => Not ready to push! (3.33 min)
1) SERIAL_RELEASE => passed: configure-only passed => Not ready to push! (3.28 min)
Other local commits for this build/test group: bdbca59
WARNING: Forced the push!