Stokhos:  Change UQ scalar types back to using execution_space.
Actually I have set things up so that a memory_space, execution_space,
or pair should work as I always extract the typedefs of these things
from the template parameter.  I prefer execution_space though to make
sure the Cijk setup is using the right exeuction space.  I also removed
the rebinding of the execution space within the scalar types when
generarting the host mirror.  This is very problematic for people if the
scalar type for their host_mirror is different for the device.