TpetraKernels: Fixed single-column element-wise multiply kernel.
The local kernel for Tpetra::MultiVector::elementWiseMultiply, for the
special case of single-column (i.e., single-vector) inputs, had a bug.
When both the ab and c coefficients were zero, the kernel did nothing,
instead of overwriting the output with zeros, as it should.  This
commit fixes that.  I also took the liberty to reformat the code.