Tpetra::MultiVector: Improved sub{View,Copy} unit test.
A few commits ago, Andrew Bradley contributed a test for exercising
MultiVector::sub{View, Copy} when some of the MultiVector's processes
have zero rows.  This commit improves the test to exercise more cases.
The main goal of the test was to catch exceptions, but I've also made
some cases actually test (a) whether the view or copy got the right
data, and (b) whether the view is a view or the copy is a (deep) copy.