Thyra::MultiVectorDefaultBase: Fixed build warning (unused typedef).
The warning referred to the following line at the end of a macro for
MultiVectorDefaultBase:

typedef THYRA_ASSERT_MV_COLS_t int

Note the lack of semicolon.  The line was probably meant to let users
follow the macro with a semicolon.  It's correct code, but it triggers
a warning for the unused typedef.

The standard idiom for writing a code block macro that users may
follow with a semicolon, is to enclose the code block in a "do { ... }
while(false)" construction, with no semicolon after the
"while(false)".  I replaced the above typedef with this idiom and
Thyra builds fine.