Tpetra::Map: Fixed build errors for new (Kokkos refactor) version.
Last year, I wrote a new version of Tpetra::Map that uses new Kokkos
for its data structures and construction.  It doesn't work right yet
with CUDA, but it's a start.  Recent changes to Kokkos' interface made
new Map not build correctly.  This commit fixes that (but not yet for
CUDA).