Tpetra: Added Doxyfile and build_docs to tpetra/doc/.
Hopefully this means that the scripts will build Tpetra's Doxygen
documentation automatically.  Those scripts look in
Trilinos/packages/*/doc for a build_docs shell script, and run it (if it
exists) in order to generate Doxygen (or other) documentation.  The lack
of this script and the corresponding Doxyfile in tpetra/doc was why
Tpetra's documentation wasn't updated for a while.  Thanks to Jim
Willenbring for clarifying this procedure.