Tpetra::CrsGraph: Fixed build error in setAllIndices().
The method accepts an Teuchos::ArrayRCP<size_t> input argument
'rowPointers'.  It's OK to make a deep copy of that array.  The
current implementation was assigning a deep copy of that array
directly to the matrix's graph's row pointers array.  The problem is
that the type of the entries of the latter may differ from size_t,
depending on the Kokkos device type.  This manifests on CUDA, where
the entries have type unsigned int, rather than unsigned long int
(size_t).  This commit aims to fix the build error.

NOTE: TpetraCore_ExpBlockCrsMatrix test still FAILS.  This has to do
with my previous commit.  Tpetra::Experimental::BlockCrsMatrix appears
to rely on getNodeRowPtrs() returning a view rather than a deep copy.