Kokkos: Added optional std::string argument to parallel_pattern
This adds an optional argument std::string to parallel_for,
parallel_reduce and parallel_scan. In the future this argument
will be used for profiling and debugging purposes. The idea is
to be able to name kernels according to their invocation point.
This is highly recommend to do for every developer since it will
help with debugging and profiling. Possible capabilities include
in build hotspot profiling, catching faults in kernels with
useful error messages and so on.

The syntax is as simple as you would expect:

parallel_for(string,policy,functor);
parallel_reduce(string,....);
parallel_scan(string,....);

string can be a literal (the functions take a const std::string&).
In debug mode (i.e. if KOKKOS_HAVE_DEBUG is defined) the functions
will print start and end messages.

Note that the cmake option for KOKKOS_HAVE_DEBUG is still missing.